… tests/fields/test_constructors.py
… tests/fields/test_instantiation.py
… tests/polys/test_constructors.py
… tests/polys/test_function.py
… tests/polys/test_operations.py
… tests/test_primes.py