- [x] Add `multiplicities=None` kwarg to `Poly.Roots()`
- [x] Add efficient calculation of `f_poly**power % g_poly`
- [x] Add irreducible check