@mhostetter 2021-04-10
Add efficient polynomial exponentiation with reduction mod another polynomial.

```python
In [1]: import galois                                                                                                          

In [2]: f = galois.Poly.Random(10, galois.GF(31)); f                                                                           
Out[2]: Poly(26x^10 + 21x^9 + 5x^8 + 24x^7 + 4x^6 + 17x^5 + 8x^4 + 11x^3 + 5x^2 + 16x + 24, GF(31))

In [3]: g = galois.Poly.Random(7, galois.GF(31)); g                                                                            
Out[3]: Poly(14x^7 + 22x^6 + 12x^5 + 8x^4 + 16x^3 + x^2 + 23x + 25, GF(31))

In [4]: f**200 % g                                                                                                             
Out[4]: Poly(19x^6 + 17x^5 + 6x^4 + 18x^3 + 28x^2 + 18x + 17, GF(31))

In [5]: galois.poly_exp_mod(f, 200, g)                                                                                         
Out[5]: Poly(19x^6 + 17x^5 + 6x^4 + 18x^3 + 28x^2 + 18x + 17, GF(31))

In [6]: %timeit f**200 % g                                                                                                     
1.25 s ± 10.3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [7]: %timeit galois.poly_exp_mod(f, 200, g)                                                                                 
39.9 ms ± 642 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```