@mhostetter 2021-04-10
Speed up irreducibility checks.

Before:

```python
In [1]: import galois                                                                                                          

In [2]: p = galois.conway_poly(7, 20)                                                                                          

In [3]: p                                                                                                                      
Out[3]: Poly(x^20 + x^12 + 6x^11 + 2x^10 + 5x^9 + 2x^8 + 3x^7 + x^6 + 3x^5 + 3x^3 + x + 3, GF(7))

In [4]: galois.is_irreducible(p)                                                                                               
Out[4]: True

In [5]: %timeit galois.is_irreducible(p)                                                                                       
3.5 s ± 226 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

After:

```python
In [1]: import galois                                                                                                          

In [2]: p = galois.conway_poly(7, 20)                                                                                          

In [3]: p                                                                                                                      
Out[3]: Poly(x^20 + x^12 + 6x^11 + 2x^10 + 5x^9 + 2x^8 + 3x^7 + x^6 + 3x^5 + 3x^3 + x + 3, GF(7))

In [4]: galois.is_irreducible(p)                                                                                               
Out[4]: True

In [5]: %timeit galois.is_irreducible(p)                                                                                       
999 ms ± 127 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```